Adding first tests to SelectAmount
