+++ client/src/Components/Flows/TippingFlow/SelectAmount/SelectAmount.test.js
@@ -4,18 +4,28 @@ import { Provider } from 'react-redux';
 import * as rtl from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
-import {SelectAmount} from './SelectAmount';
+import SelectAmount from './SelectAmount';
+import { store } from './../../../../store';
 
 afterEach(rtl.cleanup);
 
 describe('SelectAmount', () => {
 
-  it('renders', () => {
-    
+  it('renders without crashing', () => {
     const wrapper = rtl.render(
-      <div>next</div>
+      <Provider store={store} >
+        <SelectAmount />
+      </Provider>
     )
-    const element = wrapper.queryAllByText(/Next/i)
-    expect(element).toBeTruthy(); // jest matcher
+  })
+
+  it('properly renders all content, including the "Next" button', () => {
+    const wrapper = rtl.render(
+      <Provider store={store} >
+        <SelectAmount />
+      </Provider>
+    )
+    const bye = wrapper.queryByText("Next")
+    console.log(bye)
   })
 })
